pyproject.toml
[project]
name = "gigachain-toon-example"
version = "0.1.0"
description = "Minimal TOON + LangChain + GigaChat example"
requires-python = ">=3.10"
dependencies = [
  "langchain>=1.0,<2.0",
  "langchain-core>=1.0,<2.0",
  "langchain-gigachat>=0.5,<0.6",
  "python-dotenv>=1.0,<2.0",
  "python-toon>=0.1,<1.0",
  "pydantic>=2.0,<3.0",
]

[tool.uv]
package = false
